test(tooling): extract github workflow test helpers to support module
